second update for one-time pause TG2 recording
fixed gazecode further to handle TG2 recordings with one pause. As a pause is very short, there might not be a pause in the data, in the sense that the next fixation is in the next segment of scene video. Moreover videoReader (ML) seems to not find frames belonging to events that should be under the max frame number you can find with get(vidObj) but are close to that max. Look into later